refactor: move plantInPlot to its own file
